+++ packages/data-point-cache/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0](https://github.com/ViacomInc/data-point/tree/master/packages/data-point-cache/compare/data-point-cache@3.3.2...data-point-cache@4.0.0) (2019-07-24)
+
+
+### Code Refactoring
+
+* update ioredis version ([37712d1](https://github.com/ViacomInc/data-point/tree/master/packages/data-point-cache/commit/37712d1)), closes [#349](https://github.com/ViacomInc/data-point/tree/master/packages/data-point-cache/issues/349)
+
+
+### BREAKING CHANGES
+
+* new ioredis version has breaking changes
+
+
+
+
+
+## [3.3.2](https://github.com/ViacomInc/data-point/tree/master/packages/data-point-cache/compare/data-point-cache@3.3.1...data-point-cache@3.3.2) (2019-07-15)
+
+**Note:** Version bump only for package data-point-cache
+
+
+
+
+
+<a name="3.3.1"></a>
+## [3.3.1](https://github.com/ViacomInc/data-point/tree/master/packages/data-point-cache/compare/data-point-cache@3.3.1-0...data-point-cache@3.3.1) (2019-07-10)
+
+
+
+
+**Note:** Version bump only for package data-point-cache
+
 <a name="3.3.0"></a>
 # [3.3.0](https://github.com/ViacomInc/data-point/compare/data-point-cache@3.2.1-0...data-point-cache@3.3.0) (2018-12-19)